+++ booksim2/routers/bypass_router/smart_nebb/smart_nebb_vct_opt_router.cpp
@@ -88,7 +88,8 @@ namespace Booksim
 
         for (int input = 0; input < _inputs; input++) {
             if(!_flits_to_BW[input].empty()) {
-                pair<long, Flit *> elem = _flits_to_BW[input].front();
+                //BSMOD: Change time to long long
+                pair<long long, Flit *> elem = _flits_to_BW[input].front();
                 if (elem.second && elem.first == GetSimTime()) {
                     BufferWrite(input, elem.second);
                     _flits_to_BW[input].pop();
@@ -98,7 +99,8 @@ namespace Booksim
 
         for (int output = 0; output < _outputs; output++) {
             if (!_destination_queue_credits[output].empty()) {
-                pair<long, Credit *> elem =
+                //BSMOD: Change time to long long
+                pair<long long, Credit *> elem =
                     _destination_queue_credits[output].front();
                 if (elem.first >= GetSimTime()){
                     continue;
