+++ booksim2/routers/fbfcl_router.cpp
@@ -424,10 +424,11 @@ namespace Booksim
       _in_queue_flits.clear();
 
       while(!_proc_credits.empty()) {
-
-        pair<long, pair<Credit *, int> > const & item = _proc_credits.front();
-
-        long const time = item.first;
+        //BSMOD: Change time to long long
+        pair<long long, pair<Credit *, int> > const & item = _proc_credits.front();
+        
+        //BSMOD: Change time to long long
+        long long const time = item.first;
         if(GetSimTime() < time) {
           break;
         }
@@ -466,11 +467,13 @@ namespace Booksim
     {
       assert(_routing_delay);
 
-      for(deque<pair<long, pair<int, int> > >::iterator iter = _route_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<int, int> > >::iterator iter = _route_vcs.begin();
           iter != _route_vcs.end();
           ++iter) {
-        
-        long const time = iter->first;
+
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         if(time >= 0) {
           break;
         }
@@ -505,10 +508,10 @@ namespace Booksim
       assert(_routing_delay);
 
       while(!_route_vcs.empty()) {
-
-        pair<long, pair<int, int> > const & item = _route_vcs.front();
-
-        long const time = item.first;
+        //BSMOD: Change time to long long
+        pair<long long, pair<int, int> > const & item = _route_vcs.front();
+        //BSMOD: Change time to long long
+        long long const time = item.first;
         if((time < 0) || (GetSimTime() < time)) {
           break;
         }
@@ -560,11 +563,12 @@ namespace Booksim
 
       bool watched = false;
 
-      for(deque<pair<long, pair<pair<int, int>, int> > >::iterator iter = _vc_alloc_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<pair<int, int>, int> > >::iterator iter = _vc_alloc_vcs.begin();
           iter != _vc_alloc_vcs.end();
           ++iter) {
-
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         if(time >= 0) {
           break;
         }
@@ -715,12 +719,12 @@ namespace Booksim
         *gWatchOut << GetSimTime() << " | " << _vc_allocator->FullName() << " | ";
         _vc_allocator->PrintGrants( gWatchOut );
       }
-
-      for(deque<pair<long, pair<pair<int, int>, int> > >::iterator iter = _vc_alloc_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<pair<int, int>, int> > >::iterator iter = _vc_alloc_vcs.begin();
           iter != _vc_alloc_vcs.end();
           ++iter) {
-
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         if(time >= 0) {
           break;
         }
@@ -786,11 +790,12 @@ namespace Booksim
         return;
       }
 
-      for(deque<pair<long, pair<pair<int, int>, int> > >::iterator iter = _vc_alloc_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<pair<int, int>, int> > >::iterator iter = _vc_alloc_vcs.begin();
           iter != _vc_alloc_vcs.end();
           ++iter) {
-        
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         assert(time >= 0);
         if(GetSimTime() < time) {
           break;
@@ -853,10 +858,10 @@ namespace Booksim
       assert(_vc_allocator);
 
       while(!_vc_alloc_vcs.empty()) {
-
-        pair<long, pair<pair<int, int>, int> > const & item = _vc_alloc_vcs.front();
-
-        long const time = item.first;
+        //BSMOD: Change time to long long
+        pair<long long, pair<pair<int, int>, int> > const & item = _vc_alloc_vcs.front();
+        //BSMOD: Change time to long long
+        long long const time = item.first;
         if((time < 0) || (GetSimTime() < time)) {
           break;
         }
@@ -942,12 +947,12 @@ namespace Booksim
     void FBFCLRouter::_SWHoldEvaluate( )
     {
       assert(_hold_switch_for_packet);
-
-      for(deque<pair<long, pair<pair<int, int>, int> > >::iterator iter = _sw_hold_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<pair<int, int>, int> > >::iterator iter = _sw_hold_vcs.begin();
           iter != _sw_hold_vcs.end();
           ++iter) {
-        
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         if(time >= 0) {
           break;
         }
@@ -1018,10 +1023,10 @@ namespace Booksim
       assert(_hold_switch_for_packet);
 
       while(!_sw_hold_vcs.empty()) {
-        
-        pair<long, pair<pair<int, int>, int> > const & item = _sw_hold_vcs.front();
-        
-        long const time = item.first;
+        //BSMOD: Change time to long long
+        pair<long long, pair<pair<int, int>, int> > const & item = _sw_hold_vcs.front();
+        //BSMOD: Change time to long long
+        long long const time = item.first;
         if(time < 0) {
           break;
         }
@@ -1338,12 +1343,12 @@ namespace Booksim
     void FBFCLRouter::_SWAllocEvaluate( )
     {
       bool watched = false;
-
-      for(deque<pair<long, pair<pair<int, int>, int> > >::iterator iter = _sw_alloc_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<pair<int, int>, int> > >::iterator iter = _sw_alloc_vcs.begin();
           iter != _sw_alloc_vcs.end();
           ++iter) {
-
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         if(time >= 0) {
           break;
         }
@@ -1510,11 +1515,12 @@ namespace Booksim
         }
       }
       
-      for(deque<pair<long, pair<pair<int, int>, int> > >::iterator iter = _sw_alloc_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<pair<int, int>, int> > >::iterator iter = _sw_alloc_vcs.begin();
           iter != _sw_alloc_vcs.end();
           ++iter) {
-
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         if(time >= 0) {
           break;
         }
@@ -1650,12 +1656,12 @@ namespace Booksim
       if(!_speculative && (_sw_alloc_delay <= 1)) {
         return;
       }
-
-      for(deque<pair<long, pair<pair<int, int>, int> > >::iterator iter = _sw_alloc_vcs.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<pair<int, int>, int> > >::iterator iter = _sw_alloc_vcs.begin();
           iter != _sw_alloc_vcs.end();
           ++iter) {
-
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         assert(time >= 0);
         if(GetSimTime() < time) {
           break;
@@ -1855,9 +1861,10 @@ namespace Booksim
     {
       while(!_sw_alloc_vcs.empty()) {
 
-        pair<long, pair<pair<int, int>, int> > const & item = _sw_alloc_vcs.front();
-
-        long const time = item.first;
+        //BSMOD: Change time to long long
+        pair<long long, pair<pair<int, int>, int> > const & item = _sw_alloc_vcs.front();
+        //BSMOD: Change time to long long
+        long long const time = item.first;
         if((time < 0) || (GetSimTime() < time)) {
           break;
         }
@@ -2152,11 +2159,12 @@ namespace Booksim
 
     void FBFCLRouter::_SwitchEvaluate( )
     {
-      for(deque<pair<long, pair<Flit *, pair<int, int> > > >::iterator iter = _crossbar_flits.begin();
+      //BSMOD: Change time to long long
+      for(deque<pair<long long, pair<Flit *, pair<int, int> > > >::iterator iter = _crossbar_flits.begin();
           iter != _crossbar_flits.end();
           ++iter) {
-        
-        long const time = iter->first;
+        //BSMOD: Change time to long long
+        long long const time = iter->first;
         if(time >= 0) {
           break;
         }
@@ -2184,9 +2192,10 @@ namespace Booksim
     {
       while(!_crossbar_flits.empty()) {
 
-        pair<long, pair<Flit *, pair<int, int> > > const & item = _crossbar_flits.front();
-
-        long const time = item.first;
+        //BSMOD: Change time to long long
+        pair<long long, pair<Flit *, pair<int, int> > > const & item = _crossbar_flits.front();
+        //BSMOD: Change time to long long
+        long long const time = item.first;
         if((time < 0) || (GetSimTime() < time)) {
           break;
         }
